requirements.txt
Flask==2.2.2
Biopython==1.79
numpy==1.23.5
requests==2.28.1